Restructure actions
