+++ services/google-data-api/src/actions/index.js
@@ -1,0 +1,5 @@
+const youtube = require('./youtube');
+
+module.exports = {
+  youtube,
+};
