+++ services/google-data-api/src/actions/youtube/index.js
@@ -1,0 +1,7 @@
+const channelList = require('./channel-list');
+const playlistItems = require('./playlist-items');
+
+module.exports = {
+  channelList,
+  playlistItems,
+};
