+++ services/google-data-api/src/sort-object.js
@@ -1,0 +1,4 @@
+module.exports = (object = {}) => {
+  const keys = Object.keys(object).sort();
+  return keys.reduce((obj, key) => ({ ...obj, [key]: object[key] }), {});
+};
